🧩 Phase 1: FacilityManager modularization complete
✅ Successfully extracted 4 reusable components:
- CorporateInfoForm.jsx (~150 lines) - Organization type + corporate info
- FacilitySearch.jsx (~120 lines) - ACR search + results display
- SelectedFacilitiesList.jsx (~200 lines) - Facility management + stats
- ManualFacilityEntry.jsx (~100 lines) - Manual facility entry form

✅ Benefits achieved:
- Created reusable components for patient dashboard
- Improved code organization and maintainability
- Separated concerns for easier debugging
- All functionality preserved and tested working

📊 Stats: Extracted ~570 lines into modular components
🎯 Next phase: PSA Review component extraction (~400 lines)
