+++ src/components/dashboard/FacilityManager.jsx
@@ -1,3 +1,7 @@
+import CorporateInfoForm from '../facility/CorporateInfoForm';
+import FacilitySearch from '../facility/FacilitySearch';
+import SelectedFacilitiesList from '../facility/SelectedFacilitiesList';
+import ManualFacilityEntry from '../facility/ManualFacilityEntry';
 import React, { useState, useEffect, useRef } from 'react';
 import { 
   Search, Building, MapPin, Phone, Users, CheckCircle, Plus, Upload, Eye, EyeOff, 
@@ -583,133 +587,15 @@ const addManualFacility = () => {
               </div>
             </div>
 
-            {/* Enhanced Corporate Information Form */}
+            {/* Corporate Information Form - Using Extracted Component */}
             {organizationType && (
-              <div className="bg-white p-8 rounded-2xl border border-gray-200 shadow-sm">
-                <div className="flex items-center mb-6">
-                  <div className="w-12 h-12 bg-gradient-to-r from-gray-600 to-gray-700 rounded-xl flex items-center justify-center mr-4">
-                    <FileText className="h-6 w-6 text-white" />
-                  </div>
-                  <div>
-                    <h3 className="text-xl font-bold text-gray-900">
-                      {organizationType === 'single' ? 'Practice Information' : 'Corporate Information'}
-                    </h3>
-                    <p className="text-gray-600 text-sm">
-                      {organizationType === 'single' 
-                        ? 'Basic information about your imaging practice' 
-                        : 'Legal entity information for your corporate structure'
-                      }
-                    </p>
-                  </div>
-                </div>
-                
-                <div className="grid md:grid-cols-2 gap-6">
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Legal Business Name *
-                    </label>
-                    <input
-                      type="text"
-                      value={corporateInfo.legalName}
-                      onChange={(e) => setCorporateInfo({...corporateInfo, legalName: e.target.value})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="Enter legal business name"
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      DBA Name (if different)
-                    </label>
-                    <input
-                      type="text"
-                      value={corporateInfo.legalEntityName}
-                      onChange={(e) => setCorporateInfo({...corporateInfo, legalEntityName: e.target.value})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="Doing business as..."
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Federal Tax ID (EIN) *
-                    </label>
-                    <div className="relative">
-                      <input
-                        type={showEIN ? "text" : "password"}
-                        value={corporateInfo.taxId}
-                        onChange={(e) => setCorporateInfo({...corporateInfo, taxId: formatEIN(e.target.value)})}
-                        className="w-full px-4 py-3 pr-12 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                        placeholder="XX-XXXXXXX"
-                        maxLength={10}
-                      />
-                      <button
-                        type="button"
-                        onClick={() => setShowEIN(!showEIN)}
-                        className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-400 hover:text-gray-600 transition-colors"
-                      >
-                        {showEIN ? <EyeOff className="h-5 w-5" /> : <Eye className="h-5 w-5" />}
-                      </button>
-                    </div>
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Business Phone *
-                    </label>
-                    <input
-                      type="tel"
-                      value={corporateInfo.phone}
-                      onChange={(e) => setCorporateInfo({...corporateInfo, phone: formatPhoneNumber(e.target.value)})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="(XXX) XXX-XXXX"
-                      maxLength={14}
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Business Email *
-                    </label>
-                    <input
-                      type="email"
-                      value={corporateInfo.email}
-                      onChange={(e) => setCorporateInfo({...corporateInfo, email: e.target.value})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="contact@business.com"
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Website
-                    </label>
-                    <div className="relative">
-                      <Globe className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-5 w-5" />
-                      <input
-                        type="url"
-                        value={corporateInfo.website}
-                        onChange={(e) => setCorporateInfo({...corporateInfo, website: e.target.value})}
-                        className="w-full pl-10 pr-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                        placeholder="www.yourwebsite.com"
-                      />
-                    </div>
-                  </div>
-
-                  <div className="md:col-span-2">
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Authorized Signer *
-                    </label>
-                    <input
-                      type="text"
-                      value={corporateInfo.signerName}
-                      onChange={(e) => setCorporateInfo({...corporateInfo, signerName: e.target.value})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="Full name of person authorized to sign PSA"
-                    />
-                  </div>
-                </div>
-              </div>
+              <CorporateInfoForm
+                organizationType={organizationType}
+                corporateInfo={corporateInfo}
+                setCorporateInfo={setCorporateInfo}
+                showEIN={showEIN}
+                setShowEIN={setShowEIN}
+              />
             )}
           </div>
         );
@@ -717,142 +603,16 @@ const addManualFacility = () => {
         case 2:
           return (
             <div className="space-y-8">
-              {/* Enhanced Facility Search Header */}
-              <div className="bg-gradient-to-br from-emerald-50 via-green-50 to-teal-50 p-8 rounded-2xl border border-emerald-100 shadow-sm">
-                <div className="text-center mb-8">
-                  <div className="w-16 h-16 bg-gradient-to-r from-emerald-600 to-teal-600 rounded-2xl mx-auto mb-4 flex items-center justify-center shadow-lg">
-                    <Search className="h-8 w-8 text-white" />
-                  </div>
-                  <h3 className="text-2xl font-bold text-gray-900 mb-2">Add Your Imaging Facilities</h3>
-                  <p className="text-gray-600 max-w-2xl mx-auto">
-                    Search our exclusive database of 30,000+ ACR-accredited facilities or add your facilities manually. 
-                    Our intelligent system will help you build your network efficiently.
-                  </p>
-                </div>
-  
-                {/* Enhanced Search Interface */}
-                <div className="relative max-w-2xl mx-auto mb-8">
-                  <div className="relative">
-                    <Search className="absolute left-4 top-1/2 transform -translate-y-1/2 text-gray-400 h-6 w-6" />
-                    <input
-                      type="text"
-                      value={searchTerm}
-                      onChange={(e) => setSearchTerm(e.target.value)}
-                      className="w-full pl-12 pr-4 py-4 text-lg border-2 border-gray-300 rounded-2xl focus:ring-4 focus:ring-emerald-100 focus:border-emerald-500 transition-all shadow-lg hover:shadow-xl"
-                      placeholder="Search by facility name, city, state (minimum 3 characters)..."
-                    />
-                    {searchTerm && (
-                      <button
-                        onClick={() => setSearchTerm('')}
-                        className="absolute right-4 top-1/2 transform -translate-y-1/2 text-gray-400 hover:text-gray-600 transition-colors text-xl"
-                      >
-                        ×
-                      </button>
-                    )}
-                  </div>
-  
-                  {/* Search Status Messages */}
-                  {searchTerm.length > 0 && searchTerm.length < 3 && (
-                    <div className="mt-2 text-center">
-                      <p className="text-sm text-amber-600 bg-amber-50 border border-amber-200 rounded-lg px-4 py-2 inline-block">
-                        <Clock className="h-4 w-4 inline mr-1" />
-                        Type {3 - searchTerm.length} more character{3 - searchTerm.length !== 1 ? 's' : ''} to search
-                      </p>
-                    </div>
-                  )}
-  
-                  {searchTerm.length >= 3 && !isSearching && searchResults.length === 0 && (
-                    <div className="mt-2 text-center">
-                      <p className="text-sm text-gray-600 bg-gray-50 border border-gray-200 rounded-lg px-4 py-2 inline-block">
-                        <Search className="h-4 w-4 inline mr-1" />
-                        No facilities found for "{searchTerm}"
-                      </p>
-                    </div>
-                  )}
-  
-                  {/* Enhanced Search Results */}
-                  {(searchResults.length > 0 || isSearching) && (
-                    <div className="absolute z-20 w-full mt-4 bg-white border-2 border-gray-200 rounded-2xl shadow-2xl max-h-96 overflow-y-auto">
-                      {isSearching ? (
-                        <div className="p-8 text-center">
-                          <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-emerald-500 mx-auto mb-4"></div>
-                          <p className="text-gray-600 font-medium">Searching ACR database...</p>
-                          <p className="text-sm text-gray-500 mt-1">Finding matches for "{searchTerm}"</p>
-                        </div>
-                      ) : (
-                        <div className="p-2">
-                          {/* Results Header */}
-                          <div className="px-4 py-2 border-b border-gray-100 bg-gray-50 rounded-t-xl">
-                            <p className="text-sm font-medium text-gray-700">
-                              Found {searchResults.length} facilities matching "{searchTerm}"
-                            </p>
-                          </div>
-                          
-                          {/* Results List */}
-                          {searchResults.map((facility) => (
-                            <div
-                              key={facility.id}
-                              onClick={() => selectFacility(facility)}
-                              className="p-4 hover:bg-emerald-50 cursor-pointer border-b border-gray-100 last:border-b-0 rounded-xl transition-all group"
-                            >
-                              <div className="flex items-start justify-between">
-                                <div className="flex-1">
-                                  <div className="flex items-center mb-2">
-                                    <h4 className="font-bold text-gray-900 group-hover:text-emerald-700 transition-colors">
-                                      {facility.name}
-                                    </h4>
-                                    {facility.accredited && (
-                                      <div className="ml-2 flex items-center">
-                                        <CheckCircle className="h-4 w-4 text-emerald-500" />
-                                        <span className="ml-1 text-xs text-emerald-600 font-medium">ACR</span>
-                                      </div>
-                                    )}
-                                  </div>
-                                  <p className="text-sm text-gray-600 flex items-center mb-1">
-                                    <MapPin className="h-4 w-4 mr-1 text-gray-400" />
-                                    {facility.address}, {facility.city}, {facility.state} {facility.zip}
-                                  </p>
-                                  <div className="flex items-center space-x-4 text-xs">
-                                    <div className="flex items-center">
-                                      <Camera className="h-3 w-3 mr-1 text-gray-400" />
-                                      <span className="text-gray-500">
-                                        {facility.modalities ? facility.modalities.join(', ') : facility.modality}
-                                      </span>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className="ml-4 flex items-center">
-                                  {selectedFacilities.find(f => f.id === facility.id) ? (
-                                    <div className="flex items-center text-green-600">
-                                      <CheckCircle className="h-5 w-5 mr-1" />
-                                      <span className="text-sm font-medium">Added</span>
-                                    </div>
-                                  ) : (
-                                    <div className="flex items-center text-emerald-600 group-hover:text-emerald-700">
-                                      <ArrowRight className="h-5 w-5" />
-                                    </div>
-                                  )}
-                                </div>
-                              </div>
-                            </div>
-                          ))}
-                        </div>
-                      )}
-                    </div>
-                  )}
-                </div>
-  
-                {/* Enhanced Action Buttons */}
-                <div className="flex items-center justify-center space-x-6">
-                  <button
-                    onClick={() => setShowManualEntry(!showManualEntry)}
-                    className="flex items-center px-6 py-3 bg-white border-2 border-emerald-200 text-emerald-700 rounded-xl hover:bg-emerald-50 hover:border-emerald-300 transition-all font-medium shadow-md hover:shadow-lg"
-                  >
-                    <Plus className="h-5 w-5 mr-2" />
-                    Add Manually
-                  </button>
-                </div>
-              </div>
+              {/* Facility Search - Using Extracted Component */}
+<FacilitySearch
+  searchTerm={searchTerm}
+  setSearchTerm={setSearchTerm}
+  searchResults={searchResults}
+  isSearching={isSearching}
+  selectedFacilities={selectedFacilities}
+  onFacilitySelect={selectFacility}
+  onShowManualEntry={setShowManualEntry}
+/>
   
               {/* Exhibit B Context Section */}
               <div className="bg-blue-50 border border-blue-200 rounded-2xl p-6 mb-8">
@@ -886,353 +646,28 @@ const addManualFacility = () => {
                 </div>
               </div>
 
-            {/* Enhanced Manual Entry Form */}
+            {/* Manual Entry Form - Using Extracted Component */}
             {showManualEntry && (
-              <div className="bg-white p-8 rounded-2xl border border-gray-200 shadow-lg">
-                <div className="flex items-center justify-between mb-6">
-                  <div className="flex items-center">
-                    <div className="w-12 h-12 bg-gradient-to-r from-blue-600 to-indigo-600 rounded-xl flex items-center justify-center mr-4">
-                      <Edit3 className="h-6 w-6 text-white" />
-                    </div>
-                    <div>
-                      <h4 className="text-xl font-bold text-gray-900">Add Facility Manually</h4>
-                      <p className="text-gray-600 text-sm">Enter facility details for non-ACR facilities</p>
-                    </div>
-                  </div>
-                  <button
-                    onClick={() => setShowManualEntry(false)}
-                    className="text-gray-400 hover:text-gray-600 transition-colors text-xl"
-                  >
-                    ×
-                  </button>
-                </div>
-                
-                <div className="grid md:grid-cols-2 gap-6">
-                  <div className="md:col-span-2">
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Facility Name *
-                    </label>
-                    <input
-                      type="text"
-                      value={manualFacility.name}
-                      onChange={(e) => setManualFacility({...manualFacility, name: e.target.value})}
-                      className={`w-full px-4 py-3 border rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md ${errors.name ? 'border-red-500' : 'border-gray-300'}`}
-                      placeholder="Enter facility name"
-                    />
-                    {errors.name && <p className="text-red-500 text-sm mt-1">{errors.name}</p>}
-                  </div>
-
-                  <div className="md:col-span-2">
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Address *
-                    </label>
-                    <input
-                      type="text"
-                      value={manualFacility.address}
-                      onChange={(e) => setManualFacility({...manualFacility, address: e.target.value})}
-                      className={`w-full px-4 py-3 border rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md ${errors.address ? 'border-red-500' : 'border-gray-300'}`}
-                      placeholder="Street address"
-                    />
-                    {errors.address && <p className="text-red-500 text-sm mt-1">{errors.address}</p>}
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      City *
-                    </label>
-                    <input
-                      type="text"
-                      value={manualFacility.city}
-                      onChange={(e) => setManualFacility({...manualFacility, city: e.target.value})}
-                      className={`w-full px-4 py-3 border rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md ${errors.city ? 'border-red-500' : 'border-gray-300'}`}
-                      placeholder="City"
-                    />
-                    {errors.city && <p className="text-red-500 text-sm mt-1">{errors.city}</p>}
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      State *
-                    </label>
-                    <select
-                      value={manualFacility.state}
-                      onChange={(e) => setManualFacility({...manualFacility, state: e.target.value})}
-                      className={`w-full px-4 py-3 border rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md ${errors.state ? 'border-red-500' : 'border-gray-300'}`}
-                    >
-                      <option value="">Select state</option>
-                      <option value="FL">Florida</option>
-                      <option value="GA">Georgia</option>
-                      <option value="AL">Alabama</option>
-                      <option value="SC">South Carolina</option>
-                      <option value="NC">North Carolina</option>
-                    </select>
-                    {errors.state && <p className="text-red-500 text-sm mt-1">{errors.state}</p>}
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      ZIP Code
-                    </label>
-                    <input
-                      type="text"
-                      value={manualFacility.zip}
-                      onChange={(e) => setManualFacility({...manualFacility, zip: e.target.value})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="ZIP code"
-                      maxLength={5}
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-semibold text-gray-700 mb-2">
-                      Phone Number
-                    </label>
-                    <input
-                      type="tel"
-                      value={manualFacility.phone}
-                      onChange={(e) => setManualFacility({...manualFacility, phone: formatPhoneNumber(e.target.value)})}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all shadow-sm hover:shadow-md"
-                      placeholder="(XXX) XXX-XXXX"
-                      maxLength={14}
-                    />
-                  </div>
-                </div>
-
-                <div className="flex justify-end space-x-4 mt-8">
-                  <button
-                    onClick={() => setShowManualEntry(false)}
-                    className="px-6 py-3 text-gray-700 border-2 border-gray-300 rounded-xl hover:bg-gray-50 hover:border-gray-400 transition-all font-medium"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    onClick={addManualFacility}
-                    className="px-6 py-3 bg-gradient-to-r from-blue-600 to-indigo-600 text-white rounded-xl hover:from-blue-700 hover:to-indigo-700 transition-all font-medium shadow-lg hover:shadow-xl"
-                  >
-                    Add Facility
-                  </button>
-                </div>
-              </div>
+              <ManualFacilityEntry
+                manualFacility={manualFacility}
+                setManualFacility={setManualFacility}
+                errors={errors}
+                onAddFacility={addManualFacility}
+                onCancel={() => setShowManualEntry(false)}
+              />
             )}
 
 
-    {/* Enhanced Selected Facilities Management */}
-{selectedFacilities.length > 0 && (
-  <div className="bg-white p-8 rounded-2xl border border-gray-200 shadow-lg">
-    <div className="flex items-center justify-between mb-6">
-      <div className="flex items-center">
-        <div className="w-12 h-12 bg-gradient-to-r from-green-600 to-emerald-600 rounded-xl flex items-center justify-center mr-4">
-          <Shield className="h-6 w-6 text-white" />
-        </div>
-        <div>
-          <h4 className="text-xl font-bold text-gray-900">
-            Selected Facilities ({selectedFacilities.length})
-          </h4>
-          <p className="text-gray-600 text-sm">
-            These facilities will be included in your PSA as Exhibit B
-          </p>
-        </div>
-      </div>
-      
-      <div className="flex items-center space-x-4">
-        {/* Facility Count Badge */}
-        <div className="bg-blue-100 text-blue-800 px-4 py-2 rounded-lg font-semibold text-sm">
-          {selectedFacilities.length} {selectedFacilities.length === 1 ? 'Facility' : 'Facilities'} Selected
-        </div>
-        
-        {/* Filter Options */}
-        <div className="flex items-center space-x-2">
-          <Filter className="h-5 w-5 text-gray-400" />
-          <select
-            value={filterBy}
-            onChange={(e) => setFilterBy(e.target.value)}
-            className="px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent text-sm"
-          >
-            <option value="all">All Facilities</option>
-            <option value="acr">ACR Accredited</option>
-            <option value="manual">Manual Entry</option>
-            <option value="primary">Primary Only</option>
-          </select>
-        </div>
-      </div>
-    </div>
-
-    {/* Facilities Summary Stats */}
-    <div className="grid grid-cols-1 md:grid-cols-4 gap-4 mb-6">
-      <div className="bg-green-50 p-4 rounded-xl border border-green-200">
-        <div className="text-center">
-          <div className="text-2xl font-bold text-green-700">{selectedFacilities.length}</div>
-          <div className="text-sm text-green-600">Total Facilities</div>
-        </div>
-      </div>
-      <div className="bg-blue-50 p-4 rounded-xl border border-blue-200">
-        <div className="text-center">
-          <div className="text-2xl font-bold text-blue-700">
-            {selectedFacilities.filter(f => f.acrVerified).length}
-          </div>
-          <div className="text-sm text-blue-600">ACR Accredited</div>
-        </div>
-      </div>
-      <div className="bg-purple-50 p-4 rounded-xl border border-purple-200">
-        <div className="text-center">
-          <div className="text-2xl font-bold text-purple-700">
-            {selectedFacilities.filter(f => f.isManualEntry).length}
-          </div>
-          <div className="text-sm text-purple-600">Manual Entries</div>
-        </div>
-      </div>
-      <div className="bg-orange-50 p-4 rounded-xl border border-orange-200">
-        <div className="text-center">
-          <div className="text-2xl font-bold text-orange-700">
-            {selectedFacilities.filter(f => f.isPrimary).length}
-          </div>
-          <div className="text-sm text-orange-600">Primary Facility</div>
-        </div>
-      </div>
-    </div>
-    
-    {/* Enhanced Facilities Grid */}
-    <div className="grid md:grid-cols-2 gap-6">
-      {filteredFacilities.map((facility) => (
-        <div
-          key={facility.id}
-          className={`group relative p-6 rounded-2xl border-2 transition-all duration-300 hover:shadow-lg ${
-            facility.isPrimary 
-              ? 'border-green-500 bg-gradient-to-br from-green-50 to-emerald-50 shadow-md' 
-              : 'border-gray-200 hover:border-gray-300'
-          }`}
-        >
-          {/* Facility Header */}
-          <div className="flex items-start justify-between mb-4">
-            <div className="flex-1">
-              <div className="flex items-center mb-2">
-                <h5 className="font-bold text-gray-900 text-lg">{facility.name}</h5>
-                
-                {/* Status Badges */}
-                <div className="ml-2 flex items-center space-x-2">
-                  {facility.acrVerified && (
-                    <div className="flex items-center bg-green-100 text-green-800 px-2 py-1 rounded-full text-xs font-medium">
-                      <CheckCircle className="h-3 w-3 mr-1" />
-                      ACR
-                    </div>
-                  )}
-                  {facility.isPrimary && (
-                    <span className="bg-blue-100 text-blue-800 px-2 py-1 rounded-full text-xs font-bold">
-                      PRIMARY
-                    </span>
-                  )}
-                  {facility.isManualEntry && (
-                    <span className="bg-orange-100 text-orange-800 px-2 py-1 rounded-full text-xs font-medium">
-                      MANUAL
-                    </span>
-                  )}
-                </div>
-              </div>
-              
-              {/* Facility Details */}
-              <div className="space-y-2">
-                <p className="text-sm text-gray-600 flex items-center">
-                  <MapPin className="h-4 w-4 mr-2 text-gray-400" />
-                  {facility.address}, {facility.city}, {facility.state} {facility.zip}
-                </p>
-                
-                {facility.phone && (
-                  <p className="text-sm text-gray-600 flex items-center">
-                    <Phone className="h-4 w-4 mr-2 text-gray-400" />
-                    {facility.phone}
-                  </p>
-                )}
-
-                {/* Added Date */}
-                {facility.addedDate && (
-                  <p className="text-xs text-gray-500 flex items-center">
-                    <Calendar className="h-3 w-3 mr-1" />
-                    Added {new Date(facility.addedDate).toLocaleDateString()}
-                  </p>
-                )}
-              </div>
-              
-              {/* Modalities Tags */}
-              {facility.modalities && facility.modalities.length > 0 && (
-                <div className="mt-3 flex flex-wrap gap-2">
-                  {facility.modalities.slice(0, 3).map((modality) => (
-                    <span key={modality} className="px-2 py-1 bg-blue-100 text-blue-700 text-xs font-medium rounded-full">
-                      {modality}
-                    </span>
-                  ))}
-                  {facility.modalities.length > 3 && (
-                    <span className="px-2 py-1 bg-gray-100 text-gray-600 text-xs font-medium rounded-full">
-                      +{facility.modalities.length - 3} more
-                    </span>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-          
-          {/* Facility Actions */}
-          <div className="flex items-center justify-between pt-4 border-t border-gray-200">
-            <div className="flex items-center space-x-3">
-              {!facility.isPrimary && selectedFacilities.length > 1 && (
-                <button
-                  onClick={() => setPrimaryFacility(facility.id)}
-                  className="text-sm text-blue-600 hover:text-blue-700 font-medium transition-colors flex items-center"
-                >
-                  <Star className="h-4 w-4 mr-1" />
-                  Set as Primary
-                </button>
-              )}
-            </div>
-            
-            <button
-              onClick={() => removeFacility(facility.id)}
-              className="text-sm text-red-600 hover:text-red-700 font-medium transition-colors flex items-center group"
-              title="Remove from agreement"
-            >
-              <Trash2 className="h-4 w-4 mr-1 group-hover:animate-pulse" />
-              Remove
-            </button>
-          </div>
-        </div>
-      ))}
-    </div>
-
-    {/* Agreement Summary */}
-    <div className="mt-6 p-6 bg-gray-50 rounded-xl border border-gray-200">
-      <h5 className="font-semibold text-gray-900 mb-3">Agreement Summary</h5>
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-4 text-sm">
-        <div>
-          <span className="text-gray-600">Total Facilities:</span>
-          <span className="ml-2 font-semibold text-gray-900">{selectedFacilities.length}</span>
-        </div>
-        <div>
-          <span className="text-gray-600">Primary Facility:</span>
-          <span className="ml-2 font-semibold text-gray-900">
-            {selectedFacilities.find(f => f.isPrimary)?.name || 'None selected'}
-          </span>
-        </div>
-        <div>
-          <span className="text-gray-600">Agreement Type:</span>
-          <span className="ml-2 font-semibold text-gray-900">
-            {selectedFacilities.length === 1 ? 'Single Facility' : 'Multi-Facility'}
-          </span>
-        </div>
-      </div>
-      
-      {/* Next Steps */}
-      <div className="mt-4 p-4 bg-blue-50 border border-blue-200 rounded-lg">
-        <div className="flex items-center">
-          <ArrowRight className="h-5 w-5 text-blue-600 mr-2" />
-          <span className="text-sm font-medium text-blue-900">
-            Next: These {selectedFacilities.length} facilities will be automatically added to your PSA as Exhibit B
-          </span>
-        </div>
-      </div>
-    </div>
-  </div>
-)}
-          </div>
-        );
+            {/* Selected Facilities List - Using Extracted Component */}
+            <SelectedFacilitiesList
+              selectedFacilities={selectedFacilities}
+              filterBy={filterBy}
+              setFilterBy={setFilterBy}
+              onSetPrimary={setPrimaryFacility}
+              onRemoveFacility={removeFacility}
+            />
+                  </div>
+                );
 
         case 3:
         return (
@@ -1872,42 +1307,63 @@ const addManualFacility = () => {
             
             <button
               onClick={async () => {
-                if (selectedFacilities.length > 0) {
-                  // Save facilities before navigating
-                  try {
-                    setIsSaving(true);
-                    await saveProgress(); // This should save the facilities
-                    
-                    // Then determine where to go next
-                    const marketEducationCompleted = localStorage.getItem('market_education_completed');
-                    const contractTermsAccepted = localStorage.getItem('contract_terms_accepted');
-                    const psaConfirmationCompleted = localStorage.getItem('psa_confirmation_completed');
-                    
-                    console.log('Debug - localStorage status:');
-                    console.log('market_education_completed:', marketEducationCompleted);
-                    console.log('contract_terms_accepted:', contractTermsAccepted);
-                    console.log('psa_confirmation_completed:', psaConfirmationCompleted);
-                    
-                    if (!marketEducationCompleted) {
-                      console.log('Going to market education');
-                      window.location.href = '/dashboard/contract/market-education';
-                    } else if (!contractTermsAccepted) {
-                      console.log('Going to terms');
-                      window.location.href = '/dashboard/contract/terms';
-                    } else if (!psaConfirmationCompleted) {
-                      console.log('Going to confirmation');
-                      window.location.href = '/dashboard/contract/confirmation';
-                    } else {
-                      console.log('Going to PSA signing');
-                      window.location.href = '/dashboard/onboarding/psa';
-                    }
-                  } catch (error) {
-                    console.error('Error saving facilities:', error);
-                    alert('Error saving facilities. Please try again.');
-                  } finally {
-                    setIsSaving(false);
+                // Step 1: Always go to Step 2
+                if (currentStep === 1) {
+                  if (canProceedToNextStep()) {
+                    setCurrentStep(2);
                   }
-                } else {
+                }
+                // Step 2: Handle facilities and navigation
+                else if (currentStep === 2) {
+                  if (selectedFacilities.length > 0) {
+                    try {
+                      setIsSaving(true);
+                      await saveProgress();
+                      
+                      // Check if user came to edit facilities
+                      const urlParams = new URLSearchParams(window.location.search);
+                      const returnTo = urlParams.get('returnTo');
+                      
+                      if (returnTo) {
+                        // User came back to edit, return them to where they were
+                        window.location.href = returnTo;
+                      } else {
+                        // Normal flow - check localStorage flags
+                        const marketEducationCompleted = localStorage.getItem('market_education_completed');
+                        const contractTermsAccepted = localStorage.getItem('contract_terms_accepted');
+                        const psaConfirmationCompleted = localStorage.getItem('psa_confirmation_completed');
+                        
+                        console.log('Debug - localStorage status:');
+                        console.log('market_education_completed:', marketEducationCompleted);
+                        console.log('contract_terms_accepted:', contractTermsAccepted);
+                        console.log('psa_confirmation_completed:', psaConfirmationCompleted);
+                        
+                        if (!marketEducationCompleted) {
+                          console.log('Going to market education');
+                          window.location.href = '/dashboard/contract/market-education';
+                        } else if (!contractTermsAccepted) {
+                          console.log('Going to terms');
+                          window.location.href = '/dashboard/contract/terms';
+                        } else if (!psaConfirmationCompleted) {
+                          console.log('Going to confirmation');
+                          window.location.href = '/dashboard/contract/confirmation';
+                        } else {
+                          console.log('Going to PSA signing');
+                          window.location.href = '/dashboard/onboarding/psa';
+                        }
+                      }
+                    } catch (error) {
+                      console.error('Error saving facilities:', error);
+                      alert('Error saving facilities. Please try again.');
+                    } finally {
+                      setIsSaving(false);
+                    }
+                  } else {
+                    alert('Please select at least one facility before continuing.');
+                  }
+                }
+                // Other steps: normal progression
+                else {
                   setCurrentStep(Math.min(5, currentStep + 1));
                 }
               }}
