+++ src/pages/dashboard/contract/market-education.astro
@@ -8,6 +8,29 @@ import OnboardingStepper from "../../../components/onboarding/OnboardingStepper.
     <!-- Step Progress Indicator -->
     <OnboardingStepper currentStep={2} client:load />
 
+    <!-- Edit Facilities Link -->
+    <div class="text-center">
+      <a
+        href="/dashboard/onboarding/facilities?step=2&returnTo=/dashboard/contract/market-education"
+        class="text-blue-600 hover:text-blue-700 text-sm font-medium inline-flex items-center gap-2 transition-colors"
+      >
+        <svg
+          class="w-4 h-4"
+          fill="none"
+          stroke="currentColor"
+          viewBox="0 0 24 24"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            stroke-width="2"
+            d="M15.232 5.232l3.536 3.536m-2.036-5.036a2.5 2.5 0 113.536 3.536L6.5 21.036H3v-3.572L16.732 3.732z"
+          ></path>
+        </svg>
+        Edit Facility Selection
+      </a>
+    </div>
+
     <!-- Strategic Header -->
     <div class="text-center">
       <h1 class="text-3xl font-bold text-gray-900">
