+++ structure.txt
@@ -1,0 +1,273 @@
+src
+├── components
+│   ├── AboutHero.astro
+│   ├── AboutVision.astro
+│   ├── AOSInit.astro
+│   ├── BackgroundPattern.astro
+│   ├── BookingFlow.jsx
+│   ├── Breadcrumb.astro
+│   ├── CallToAction.astro
+│   ├── CenterCard.astro
+│   ├── ChatbotToggle.astro
+│   ├── ClarityMark.astro
+│   ├── ClinicalCredentialsGrid.astro
+│   ├── ClinicalIntroWithImage.astro
+│   ├── ClinicalTrustSection.astro
+│   ├── CondensedProblemSolution.astro
+│   ├── CondensedTrustIndicators.astro
+│   ├── ConsultationForm.astro
+│   ├── CoreBenefits.astro
+│   ├── corporate-dashboard
+│   │   └── CorporateDashboardSystem.jsx
+│   ├── Counter.tsx
+│   ├── dashboard
+│   │   ├── AuthenticationDebugger.jsx
+│   │   ├── CorporateDashboardLayout.astro
+│   │   ├── DashboardLayout.astro
+│   │   ├── EnhancedFacilityManager.jsx
+│   │   ├── EnhancedOnboardingSystem.jsx
+│   │   ├── EnhancedPSASystem.jsx
+│   │   ├── EnhancedSettingsSystem.jsx
+│   │   ├── EnterpriseFacilityManager.jsx
+│   │   ├── EnterpriseFacilityManager.jsx.backup
+│   │   ├── FacilityManager.jsx
+│   │   ├── FacilityManager.jsx.backup
+│   │   ├── IntegrationVerifier.jsx
+│   │   ├── OnboardingDetector.jsx
+│   │   ├── PSADocumentEmbed.jsx
+│   │   ├── PSAProgressWrapper.jsx
+│   │   ├── PSASigningSystemAPI.jsx
+│   │   ├── PSASigningSystemEmbedded.jsx
+│   │   ├── SettingsProfile.jsx
+│   │   ├── SignOutTester.jsx
+│   │   ├── SkeletonAnalyticsSystem.jsx
+│   │   ├── SkeletonAppointmentsSystem.jsx
+│   │   ├── SkeletonFacilitySystem.jsx
+│   │   ├── SkeletonOnboardingSystem.jsx
+│   │   ├── SkeletonProviderDashboardSystem.jsx
+│   │   ├── SkeletonPSASystem.jsx
+│   │   ├── SkeletonReportsSystem.jsx
+│   │   ├── SkeletonSettingsSystem.jsx
+│   │   ├── UnifiedSettingsSystem.jsx
+│   │   └── UserDataFlowTester.jsx
+│   ├── DashboardNavbar.astro
+│   ├── employer
+│   │   ├── HeroEmployer.astro
+│   │   └── PartnerMfgGrid.astro
+│   ├── EmployerConsultationForm.astro
+│   ├── FadeIn.astro
+│   ├── FAQHome.astro
+│   ├── FAQPreview.astro
+│   ├── Footer.astro
+│   ├── Header.astro.backup
+│   ├── HealthcareCoverageCrisisGraph.jsx
+│   ├── HealthcareCoverageCrisisGraph.tsx
+│   ├── HeroPatientIntro.astro
+│   ├── HeroSection.astro
+│   ├── HomeProblemSolution.astro
+│   ├── HomeTrustIndicators.astro
+│   ├── HowItWorks.astro
+│   ├── HowItWorksSteps.astro
+│   ├── imaging
+│   │   ├── HeroImagingCenter.astro
+│   │   └── NetworkMap.astro
+│   ├── LoginDropdown.astro
+│   ├── LogooutButton.jsx
+│   ├── ManufacturerCredibility.astro
+│   ├── Map.astro
+│   ├── MfgTabs.astro
+│   ├── MobileMenu.jsx
+│   ├── NetworkApplicationForm.astro
+│   ├── NetworkMap.astro
+│   ├── News.astro
+│   ├── onboarding
+│   │   ├── FounderWelcomeMessage.astro
+│   │   └── OnboardingStepper.jsx
+│   ├── OurBusinessModel.astro
+│   ├── OurModel.astro
+│   ├── partner-old
+│   │   ├── FAQCard.astro
+│   │   ├── HeroSmartSplit.astro
+│   │   ├── HowItWorksTimeLine.astro
+│   │   ├── HowItWorksVisualGrid.astro
+│   │   ├── IndependentInvitation.astro
+│   │   ├── MarketOpportunity.astro
+│   │   ├── PartnerMfgGrid.astro
+│   │   └── StatCard.astro
+│   ├── patient-dashboard
+│   │   ├── PatientDashboardLayout.astro
+│   │   ├── SkeletonAppointmentsSystem.jsx
+│   │   ├── SkeletonBillingSystem.jsx
+│   │   ├── SkeletonBookingSystem.jsx
+│   │   ├── SkeletonDashboardSystem.jsx
+│   │   ├── SkeletonEducationSystem.jsx
+│   │   ├── SkeletonProfileSystem.jsx
+│   │   ├── SkeletonReferralsSystem.jsx
+│   │   ├── SkeletonReportsSystem.jsx
+│   │   └── SkeletonSupportSystem.jsx
+│   ├── PatientHeader.jsx
+│   ├── PlaceholderPage.astro
+│   ├── PrimaryButton.astro
+│   ├── ProblemStatement.astro
+│   ├── PromiseBanner.astro
+│   ├── psa
+│   │   └── EnhancedPSAComponent.jsx
+│   ├── SecondaryButton.astro
+│   ├── SectionDivider.astro
+│   ├── SectionHeading.astro
+│   ├── shared
+│   │   ├── DataSyncProvider.jsx
+│   │   └── ProgressiveDataTracker.jsx
+│   ├── SmartMatch.astro
+│   ├── SmartMatchData.ts
+│   ├── SmartMatchShowcase.astro
+│   ├── StatsCard.astro
+│   ├── stores
+│   │   └── authStore.js
+│   ├── TeamCommitment.astro
+│   ├── TrustBadgesBar.astro
+│   ├── ui
+│   │   ├── Accordion.astro
+│   │   ├── alert.tsx
+│   │   ├── badge.tsx
+│   │   ├── button.tsx
+│   │   ├── card.tsx
+│   │   └── progress.tsx
+│   ├── UserContext.jsx
+│   ├── UserContextWrapper.astro
+│   ├── UtilityHeader1.astro
+│   ├── UtilityHeader.astro
+│   └── WhyItCostsLess.astro
+├── data
+│   ├── centers.json
+│   ├── news.json
+│   └── studies.json
+├── env.d.ts
+├── hooks
+│   ├── useAutoSave.ts
+│   ├── useUserDataFlow.ts
+│   └── useUSRadAuth.js
+├── layouts
+│   ├── MainLayout.astro
+│   ├── MainPatientLayout.astro
+│   ├── PartnerLayout.astro
+│   └── PartnerPageLayout.astro
+├── lib
+│   ├── db.ts
+│   ├── facilityManager.js
+│   ├── getServerUser.js
+│   ├── supabase.js
+│   └── utils.ts
+├── motion
+│   └── motion-presets.js
+├── pages
+│   ├── about.astro
+│   ├── accessibility.astro
+│   ├── api
+│   │   ├── check-psa-status.ts
+│   │   ├── docuseal
+│   │   ├── docuseal-webhook.js
+│   │   ├── facilities
+│   │   ├── hello.ts
+│   │   ├── provider
+│   │   └── update-psa-status.js
+│   ├── auth
+│   │   ├── reset-password.astro
+│   │   └── verified.astro
+│   ├── background-test.astro
+│   ├── book.astro
+│   ├── careers.astro
+│   ├── co-founder.astro
+│   ├── contact.astro
+│   ├── corporate-dashboard
+│   │   ├── components
+│   │   └── index.astro
+│   ├── dashboard
+│   │   ├── analytics.astro
+│   │   ├── appointments.astro
+│   │   ├── auth-debug.astro
+│   │   ├── center.astro
+│   │   ├── components
+│   │   ├── contract
+│   │   ├── index.astro
+│   │   ├── onboarding
+│   │   ├── patient.astro
+│   │   ├── psa
+│   │   ├── referral.astro
+│   │   ├── reports.astro
+│   │   ├── settings
+│   │   └── settings.astro
+│   ├── employer
+│   │   └── schedule.astro
+│   ├── employer.astro
+│   ├── faq4.astro
+│   ├── faq.astro
+│   ├── form-submitted.astro
+│   ├── founder.astro
+│   ├── how-it-works.astro
+│   ├── imaging-center
+│   │   ├── apply.astro
+│   │   ├── benefits.astro
+│   │   ├── calculator.astro
+│   │   ├── coming-soon.astro
+│   │   ├── experience.astro
+│   │   ├── faq.astro
+│   │   ├── implementation.astro
+│   │   ├── index.astro
+│   │   ├── model.astro
+│   │   ├── schedule.astro
+│   │   ├── signup.astro
+│   │   └── support.astro
+│   ├── index.astro
+│   ├── investor.astro
+│   ├── locations.astro
+│   ├── login
+│   │   ├── imaginglogin.astro
+│   │   ├── index.astro
+│   │   ├── patientlogin.astro
+│   │   └── referrallogin.astro
+│   ├── managed-care.astro
+│   ├── news.astro
+│   ├── our-approach.astro
+│   ├── partner1.astro
+│   ├── partner2.astro
+│   ├── partner-experience.astro
+│   ├── patient-dashboard
+│   │   ├── appointments
+│   │   ├── billing
+│   │   ├── booking
+│   │   ├── components
+│   │   ├── education
+│   │   ├── index.astro
+│   │   ├── profile
+│   │   ├── referrals
+│   │   ├── reports
+│   │   └── support
+│   ├── press-kit.html
+│   ├── privacy-policy.astro
+│   ├── provider-signup.astro
+│   ├── schedule.astro
+│   ├── search.astro
+│   ├── technology.astro
+│   ├── terms-of-service.astro
+│   └── test.api.ts
+├── scripts
+│   └── unifiedForm.js
+├── services
+│   ├── onboardingRouter.ts
+│   └── userDataFlowManager.ts
+├── styles
+│   └── global.css
+├── test
+└── utils
+    ├── debounce.js
+    ├── domUtils.js
+    ├── formHandlers.js
+    ├── psaAutoPopulation.js
+    ├── settingsUtils.js
+    ├── supabaseHelpers.ts
+    ├── supabaseService.js
+    ├── toastService.js
+    └── validators.js
+
+50 directories, 221 files
